+++ packages/plugin-packages/mashroom-http-proxy/test/proxy/ProxyImplRequest.test.ts
@@ -81,7 +81,7 @@ describe('ProxyImplRequest', () => {
             .get('/foo')
             .reply(200, 'test response');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -102,7 +102,7 @@ describe('ProxyImplRequest', () => {
             .post('/login')
             .reply(200, 'test post response');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('POST', '{ "user": "test }');
         const res = createDummyResponse();
@@ -119,7 +119,7 @@ describe('ProxyImplRequest', () => {
             .get('/foo?q=javascript%205')
             .reply(200, 'test response');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         req.query = {
@@ -138,7 +138,7 @@ describe('ProxyImplRequest', () => {
             .get('/foo?bar=2&q=javascript%205')
             .reply(200, 'test response');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         req.query = {
@@ -153,7 +153,7 @@ describe('ProxyImplRequest', () => {
     });
 
     it('sets the correct status code if the target is not available', async () => {
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -170,7 +170,7 @@ describe('ProxyImplRequest', () => {
             .delay(3000)
             .reply(200, 'test response');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -189,7 +189,7 @@ describe('ProxyImplRequest', () => {
             .get('/foo')
             .reply(201, 'resource created');
 
-        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -248,7 +248,7 @@ describe('ProxyImplRequest', () => {
         };
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -287,7 +287,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['another-header']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -328,7 +328,7 @@ describe('ProxyImplRequest', () => {
         };
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, removeAllHeaderFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, removeAllHeaderFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -367,7 +367,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['another-header']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         req.query.foo = 'bar';
@@ -405,7 +405,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['another-header']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -445,7 +445,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['another-header']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -486,7 +486,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['x-whatever', 'some-new-header']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -531,7 +531,7 @@ describe('ProxyImplRequest', () => {
         const interceptorHandler = new InterceptorHandler(pluginRegistry);
         const headerFilter = new HttpHeaderFilter(['x-whatever', 'foo']);
 
-        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, loggerFactory);
+        const httpProxy = new ProxyImplRequest(2000, interceptorHandler, headerFilter, false, null, loggerFactory);
 
         const req = createDummyRequest('GET');
         const res = createDummyResponse();
@@ -541,5 +541,17 @@ describe('ProxyImplRequest', () => {
         expect(res.headers).toEqual({
             'x-whatever': '123'
         });
+    });
+
+    it('it rejects requests with invalid HTTP protocols', async () => {
+        const headerFilter = new HttpHeaderFilter([]);
+        const httpProxy = new ProxyImplRequest(2000, noopInterceptorHandler, headerFilter, false, null, loggerFactory);
+
+        const req = createDummyRequest('GET');
+        const res = createDummyResponse();
+
+        await httpProxy.forward(req, res, 'file:///foo/test.exe');
+
+        expect(res.statusCode).toBe(502);
     });
 });